requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
python-pptx
Pillow